+++ EcoClean/Models/ColetaTipoResiduoModel.cs
@@ -2,10 +2,10 @@
 {
     public class ColetaTipoResiduoModel
     {
-        public int ColetaId { get; set; }
+        public long ColetaId { get; set; }
         public ColetaModel Coleta { get; set; }
 
-        public int TipoResiduoId { get; set; }
+        public long TipoResiduoId { get; set; }
         public TipoResiduoModel TipoResiduo { get; set; }
     }
 }
